Moving conn related funcs to a separate file
